move the data holding area into Puppet::Type::Datacat::Common
The scope you are in during the Puppet::Type.type(:foo) block isn't a real
class, it's the top-level scope, so you're messing around with Object.

By pushing the data assembly region into Puppet::Type::Datacat::Common we avoid
the error, and should make it a little easier to follow the logic.