reproduction case for #1
If you refer to something like scope you get a deeply unfriendly
error like so:

    $ puppet apply --modulepath spec/fixtures/modules -e 'include issue1'
    Error: /Stage[main]/Issue1/Datacat[/tmp/issue1.1]/Datacat_collector[/tmp/issue1.1]: Could not evaluate: undefined local variable or method `scope' for #<Puppet_X::Richardc::Datacat_Binding:0x007f9b9e08c7f8 @data={}>

At the very least it should give a filename and linenumber rather than
just an object reference.

Maybe as it's scope we should add a more specific reminder that this
template isn't evaluated in the regular way.  Also the documentation
should stress that the template won't be run on the master so that this
isn't expected to work.  Either way it needs to be more friendly.