Move libvirt options into a group
There are a bunch of libvirt specific flags which are currently in the
default group. This is confusing to users of other hypervisors, and poor
hygiene. We should move these options into a group of their own.

from: https://blueprints.launchpad.net/nova/+spec/libvirt-opts-group

this patches still contain libvirt_use_virtio_for_bridges in the DEFAULT
section