bump head to support for managing ssh keys
This change permits provisioning ssh public and private keys as part
of creating the "nova" user.  There are no changes to the default
behavior
    of the module.  This changes introduces three new parameters:

    - nova_public_key
    - nova_private_key
    - nova_shell

    These changes are intended to permit configuring live migration via
    qemu+ssh connections to remote hypervisors.

this includes the commit that: Move libvirt options into a group

There are a bunch of libvirt specific flags which are currently in
the default group. This is confusing to users of other hypervisors, and
poor hygiene. We should move these options into a group of their own.

from: https://blueprints.launchpad.net/nova/+spec/libvirt-opts-group

this patches still contain libvirt_use_virtio_for_bridges in the
DEFAULT section