install: Use include ::apache rather than calling the class
With the way the module is currently written, it is hard to make
it colive with other modules that relies on apache. It will end up
on a Duplicate resource.

By including it, we can still let kibana3 manage webservice (vhost)
but yet colive with other modules. Configuration might be done with
hiera or another solution.