td-agent.conf: Handle both syntax for fluentd v1 and v2
Few days ago, I've submitted a PR that was adding an @ in front
of the include as indicated per the source code.
The issue is that syntax is only valid for v2 and is not backward
compatible with v1. This PR fixes that. Sorry for the inconvenience.

PR is https://github.com/mmz-srf/puppet-fluentd/pull/29