Modify how IPA configuration is detected and what is issued
Drop the "is-IPA-configured" fact and use $onlyif in the Exec
instead. This is rather simplistic but is no worse than the fact.

Drop the semaphore file for NSS-based certificates and and
File entries for OpenSSL certs and keys and replace with an
$unless call to use certmonger to determine whether a given
certificate is already being tracked. This is a lot less
error-prone.