Use array of hashes for the sensu API endpoints
This commit deprecates uchiwa::api as method to pass API endpoints
configuration to the uchiwa.conf template. An array of hashes is
passed to init.pp instead, representing the params needed to
populate the sensu API endpoints section in uchiwa.conf.
This commit may create backward compatibility issues for existing
versions of uchiwa-puppet.