Remove require relationship on python
The python package declaration was removed in 11fe6a76, so the require
relationship between python-greenlet and the python package is no
longer needed and causes errors. The python-greenlet package will have
python as a dependency if it is not already installed, so there is no
reason to make this relationship explicit.

Change-Id: I783a3a0b1d305a491e710621acd26f55b5979a90