stop puppet from breaking neutron
The neutron-ovs-cleanup service is meant to run at boot to tear down ovs
interfaces created by Neutron.  The service runs and then exits, which
means that from Puppet's perspective it is always "stopped".

By setting `ensure => started` on the neutron-ovs-cleanup service,
puppet was starting this service every time it runs, which would
completely disable Neutron networking.

This patch removes the `ensure` setting, such that the service will be
enabled and will run at boot but will not be erroneously started by
Puppet.

Conflicts:
	manifests/agents/ml2/ovs.pp

Change-Id: I623a3a0b672e00ddb8513cc346cbbfeca8fdf88b