Deprecate MidoNet's puppet-zookeeper
Several changes to deprecate the MidoNet's zookeeper module. Deric's
Zookeeper:

  https://github.com/deric/puppet-zookeeper

Is well mantained and it has much more functionality. So the commit
replaces the dependencies and sets some changes to adapt to the new
module.

I have kept the ::midonet::zookeeper manifest to maintain backwards
compatiblity with tools that use MidoNet 2015.01 version.

So the ::midonet::zookeeper manifest is just a gangway to the actual's
deric one now.

Change-Id: I6e6ebe8c9fea2e3d3195d9890b1c89362bd23d58