Filter ODL jobs based on user and process name
On my Ubuntu 14.04 based system the previous script was causing
multiple instances of ODL to spawn due to the fact that PID tracking
was searching for `karaf` process, whereas the the karaf process
spawned a `java` instance. Hence Upstart tracked a wrong PID and
assumed that ODL job is always stopped. That also caused inability to
stop the service, because PID was incorrect, so upstart was sending
SIGTERM to a wrong process.
With this change, upstart searches for a `java` process belonging to
user `odl` in order to determine the PID.