Delay ODL startup and disable console
Delay ODL startup until networking service is up, otherwise if all
interfaces on the box come up with DHCP and there is no local
definition of FQDN (as in `/etc/hosts`), ODL fails to startup.

Secondly disable the server console when starting up, otherwise it will
hijack the startup tty.